pyproject.toml
[project]
name = "ai-toolbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "langchain[openai]>=0.3.25",
    "lxml>=5.4.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
